eb-virt-ayush/BharatMenuBackend/requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.1.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
autopep8==1.7.0
billiard==4.2.0
boto3==1.28.84
botocore==1.31.84
celery==5.3.6
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==38.0.1
Django==4.1.1
django-cors-headers==3.13.0
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
django-rest-knox==4.2.0
django-storages==1.13.1
djangorestframework==3.14.0
exceptiongroup==1.2.0
frozenlist==1.3.3
gTTS==2.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
jmespath==1.0.1
kombu==5.3.4
multidict==6.0.4
openai==0.27.2
packaging==23.2
phonenumbers==8.12.56
Pillow==9.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pycparser==2.21
pycurl==7.45.2
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pysqlite-binary==0.5.1.3380300
python-dateutil==2.8.2
pytz==2022.2.1
redis==5.0.1
replicate==0.20.0
requests==2.28.2
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
toml==0.10.2
tqdm==4.65.0
twilio==8.11.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.12
yarl==1.8.2
